Implement SIGTRAP based anti-debug
Periodically signal the parent with an int3, check that the installed
SIGTRAP handler increments a global variable, if not (which should
happen if we're being debugged), kill child and exit.

Note that this requires changing the linker script/link options such
that the loader code is no longer position independent. This is due to a
quirk of how ld handles linking to a flat binary, see:
https://gcc.gnu.org/pipermail/gcc/2020-December/234571.html.