Split out trap point and function arrays
Earlier one struct function was embedded in each struct trap_point as a
matter of convenience (code was much easier to write this way). This was
both space-inefficent and led to the loader code being ugly in many
places as there were several sources of truth on the subject of function
information. This commit adds a data flexible array member to struct
trap_point_info which contains an array of trap points and functions.
Functions are referenced in each trap point by an index into the
functions array.

In addition to the advantages mentioned above, this also much more
easily permits having unique rc4 keys per function.