Omit runtime from loader when packing with -n
Modifies the loader build system to produce two loaders. One with a
runtime, and one without. The one without is used when packing with the
-n flag.

The runtime was a large amount of bulk that wasn't necessary in layer 1
only packed binaries. Furthermore, it prevented the execution of
multithreaded programs (as the runtime only supports single threaded
programs). This change allows for easy packing of eg. the Python
interpreter when used with the -n flag.