First stab at full multithreading support
This is fragile, but provides a tentative working implementation of a
multithreaded-aware runtime that can be iterated on.

The runtime code previously assumed a single thread of execution. This
assumption _greatly_ simplified things, but would cause dramatic
implosions if a binary did a fork/vfork/clone. A lot of bookkeeping has
been added to runtime.c to keep track of and synchronize multiple
threads of execution such that they're all executing decrypted code and
can make progress.

The main, and most fundamental change made to the runtime was to get rid
of the encrypted bit in struct function. This has been replaced with a
"refcount", which is bumped every time a thread of execution enters that
function and decremented every time a thread leaves the function. A
refcount of 0 thus indicates that the function is not in the call stacks
of any threads and can be encrypted. Refcounts are maintained per
*address space* (not thread). Consider a binary that does a clone
syscall with the CLONE_VM flag specified and then a fork. We now have
three threads of execution, two of which share the same address space
and one which has a separate address space (the product of the fork
call). Encrypting a function in the clone'd thread will not change
anything in the address space of the fork'd thread, thus separate
bookkeeping is required.

To maintain all this new metadata, a fully-fledged malloc implementation
has been added to loader/. This is something I'd been putting off for a
while, but is now a necessity to maintain the linked data structures
required to perform all this bookkeeping.

Upon a thread trapping into the runtime, all threads in the same address
space must be stopped to ensure they can't call a function
mid-decrypt/encrypt. This is accomplished by sending a SIGSTOP to them
all. Whenever the main runtime loop returns from its wait4 syscall due
to one of these SIGSTOP calls, it simply continues the thread and
continues, ensuring every thread will get restarted eventually.